[LLT-3324] Fix multi second stall in enabling magic dns
When running dns query handling we kept read lock during a lookup that
could query external dns servers. This caused other places that wanted
to acquire that lock with write access to be stalled for the duration of
that lookup.

After this change we limit duration of the read lock to only a clone of
Arc. The write lock code uses clone on write to reduce the time it takes
to get write access to zones.