Monitor tokio threads for being parked for a long period of time.
It is possible for tokio to get into a state where no progress is
made: https://github.com/tokio-rs/tokio/issues/4730 . This PR adds
monitoring that tries to detect such case. We don't have access to
information which threads are waiting on IO driver so this code can
produce false warnings.