Check for macos_sideload during a call to telio_new
This change makes macos_sideload mandatory on macos and removes it on
other systems. This way it's not possible to use it in code that is
meant to be os agnostic. This change also introduces new (cargo) feature
flag to control that behaviour so that it's possible to test it outside
of macOS.