Bump version of our fork of trust-dns to 0.22.1
This involves small changes to the api and migration of upstream to
tracing from log. To keep the changes in our fork we go along with that
migration and add translation from tracing events to log messages in
telio-utils that is always set up when telio is started via ffi call.