Add nat-lab tests for meshnet_id reporting via moose
One test to verify that a single node meshnet can change reported id
when a second node joins. Second one which verifies that after restarting
tcli on with the same config it reports the same meshnet_id as before
the restart. Other existing tests extended to verify (via validators) that
all meshnet_ids are the same.