Implement support for IPv6 in firewall
After this change the internal connection tracking mechanism handles
both tcp and udp packets in the exactly same way regardless of
underlying IP version.

The main implementation difference is that instead of storing IP
source/destination as u32 we know store them as an enum of u32 or u128.
All other logic remains the same.