Add telio-firewall benchmarks CI job
This job is just to keep the benchmarks from being broken. This is why
the durations are reduced - we only care that it compiles and runs
correctly. For the same reason we only run for one combination of
peers/connections - 64/10000. The timings on CI are not useful since
the machine they are run on can be arbitrarily configured and loaded.

Signed-off-by: Tomasz Kłak <tomasz.klak@nordsec.com>