Firewall benchmarks measure both ip v4 and v6 performance
Packets sent/received in benchmarks are abstracted to allow for uniform
handling of both IPv4 and IPv6 packets (in the same way as we did in the
firewall tests). This doubles number of benchmarks.

At the same time different number of packets to send/receive is removed
and only the bigger number of packets is used. This halves the number of
benchmarks which in the end results in keeping the same number of
benchmarks overall.

Signed-off-by: Tomasz Kłak <tomasz.klak@nordsec.com>