Remove nagger job
This job is right now just a dummy. It builds dylint but runs zero
checks with it. Since it didn't use specific version of dylint, it
started to fail when the dylin executable started to require rust
version newer than what we use. Instead of fixing it now, let's remove
it now. It can be readded once we have our lints open sourced and
ready to be used here.

Signed-off-by: Tomasz Kłak <tomasz.klak@nordsec.com>