Add scheduled gitlab runs
Start gitlab pipelines every 30 minutes during the night. This is very
similar to the old gitlab based scheduled pipelines. We use the new
functionality of the action to mark our scheduled pipelines. This will
be used later on to collect them and generate summaries and statistics
of the failing nat-lab tests.

The schedule is slightly different from the old gitlab one, to avoid the
crowded full hour mark.