Deduplicate peer IPs
Wireguard uAPI mandates that allowed IPs be unique for each peer,
yet we haven't made sure that is the case. We already know that
internal peers will not have IPs that overlap with each other, but
external nodes may overlap with internal and/or external nodes.
If there is overlap we simply remove the IP from the external node.