natlab: Fix test teardown
Until now, if an exception was thrown during test execution, whether
for timing out or any other reason, context managers were not able to
properly exit. When timing out from pytest-timeout, the plugin
throws a TimeOutError exception and it doesn't allow the user to
catch it. Instead it handles it internally and replaces it by a
pytest.fail() call, which in its turn raises a Fail exception.

Our actual implementations doesn't handle every kind of exception,
naturally, and whether we have an active ssh or docker connection
the teardown needs to be carefully handled so that the cleanup
takes place.

This commit adds:
  - a custom event loop, where all the tasks are manually cancelled
  when the test execution is going to stop.
  - checks if tcli process is active before cleaning up, otherwise
  the Process concrete class will take care of it during teardown.

This not the ideal solution as the cleanup we design is not really
happening, instead we cleanup by killing the processes. Nevertheless,
this fixes the actual issue where configuration leftovers are not
cleaned from the container, making other tests to fail.

With the ongoing improvement of the framework, we are going to be
able to make pytest.fixtures from setup and teardown functions,
isolating the setup/teardown phase from the test itself, avoiding
these type of issues.