Remove assertion which always fails but races with the rest of the test
The assertion which checks if the not existing executable is on the list
of running processes, would **always** fails. But it is most of the time
overriden by the subprocess creation throwing expected exception.
Sometimes the async with body ends before prallel subprocess execution
reaches the excetpion raising code and in those situation the test fails
due to the assertion failure.

Since the assertion would always fail when reached and is occasionally
reached, it is safe to remove it since it is incorrect. To make sure
that the subprocess has time to fail we sleep/wait for 1s.

This has been confirmed with additional logging and also by adding a
sleep before the `raise` in the `DockerProcess.execute` which causes
consistent test failures due to wrong exception being thrown.