Fix infinite loop
An infinite loop was created in our STUN handler because a timeout
was set too late. The timeout determines if we should try to start
a STUN session or wait, but it was set after trying to start the
STUN session. In cases where starting the STUN session failed, the
timeout would never be updated, making the code think that we
should start a STUN session, even though we just tried, leading
to an infinite loop.