Change when trigger_endpoint_candidates_discovery is called
trigger_endpoint_candidates_discovery() is responsible for
WaitingForWG->SearchingForServer transition in stun endpoint povider
state machine. Therefore this function should only be called upon
event of WG-STUN peer insertion, while previously it was called on every
wg_controller iteration which had WG-STUN peer in requested_peers as
well as an actual peer within the WG. After the changes, it is called
only when a peer is inserted. Note that whenever a server rotation
occurs, there a new call because completely new peer is created as
identified by public key.