Add handling of icmp error types to firewall
A recent PR added firewall logic for filtering ICMP packets,
but to keep the PRs simpler, that PR ignored ICMP error types.
This PR adds that error handling.

When an ICMP request fails, the error response contains the
IP packet of the request in the body of the response packet.
So in addition to the previous logic of checking that inbound
packets are direct responses to outbound packets, error responses
are also accepted if the nested IP packet matches that of an
outbound packet