Add mesh file-config <meshmap.json> command to tcli
Sometimes one needs to specify completely custom meshmap JSON to check
out various conditions manually. While we do have `mesh config <meshmap
JSON>` command it is not very convenient to use when doing manual tests.
This PR adds aditional command which behaves in a exactly the same
manner as `mesh config` behaves except for the fact that meshmap is read
from file specified as an argument instead of required to specify the
whole json as an argument to the command.