Implement TCP RST packet generation from the firewall for IPv4 connections
It's done to prevent the connections of becoming stale when the VPN server is changed.
We're using the information from the firewall to construct TCP RST packets with a proper sequence number
for all TCP connections.
The packets origin is set to the public peer's IP and the desitnation to the local client IP,
thus tricking the client into thinking the TCP connection is reset by the remote peer.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>