Implement ICMP Destination Unreachable packet generation per UDP connection from the firewall for IPv4 connections
It's done to prevent the UDP reads from becoming stale when the VPN server is changed.
We're using the information from the firewall to construct ICMP packets for all UDP connections.
The packets' origin is set to the public peer's IP and the destination to the local client IP,
thus tricking the client into thinking the destination UDP port is unreachable.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>