Add no-link detection mechanism
Add a hinting mechanism for no-link detection. If the time_since_last_rx
is greater than WG_KEEPALIVE_INTERVAL + CONFIGURABLE_RTT then a node
event with a LinkState::Down will be emitted. The event generation can
be configured using Features and the field in the event is optional.
More info about this can be found in RFC-LLT-0045.