Flush libmoose changes upon sending heartbeat event
libtelio heartbeat event is generally a once an hour occurance,
therefore it is expected that event throughput is so low, that the moose
queue will never contain more than one element. What is more, in tests
there is an expectation that after triggering the analytics, database
will contain events relatively soon. For some reason this does not
happen in about 10% of the cases. This change is an attempt to verify
whether flushing has impact on the event queueing.