Change the CPC optimization to evaluate time_since_last_rx
In general CPC needs to know when a peer is considered offline, when
online. Previously time since last WireGuard handshake has been used.
But this showed to be problematic in tests, as the shortest amount the
thershold can be lowered to is ~180s, or the maximum allowed time for WG
session to be alive. But this would force tests to wait for 180s until
they can verify, that CPC has stopped sending CMMs peers, which can
explode test-suite time.

What is more, there is a better metric, which is even more practical in
the implementation of CPC. Instead of coupling time since handshake last
handshake to the peer status (online/offline) it seems like it is better
idea to tie time since list received data. This includes handshakes
_and_ can be easily triggered by testing suites.

This PR changes CPC to start tracking the time since last recieved
packet instead of handshakes, and fixes incorrectly written
`test_direct_working_paths_with_skip_unresponsive_peers` test.