Add no-link detection mechanism
Add a hinting mechanism for no-link detection. The event generation can
be configured using Features and the field in the event is optional.
More info about this can be found in RFC-LLT-0045.

This commit also includes nat-lab tests for this detection mechanism and
a fix in is_same_event() discovered while working on it.