Fix crash in wireguard-go
Wireguard-go could crash if winrio initization fails. That is because in
that case StdNetBind is returned from NewDefaultBind function instead of
WinRingBind. Later this object was casted to the WinRingBind in order to
call OpenOnLocalhost method. This commit adds new method OpenOnLocalhost
to our interface FullBind and implements is for both WinRingBind and
StdNetBind. It requires us to copy StdNetBind implementation for the
upstream.