Fix CPC unresponsive threshold
CPC implemented an optimization to avoid sending CMMs and pings to nodes
which are offline. The "offline" in the context of this optimization is
determined by the time since last handshake. Wireguard attempts new
handshake every 120-180s, but somehow the value in the tests for CPC
"offline" threshold has been set to 10s. This means, that every time,
when direct connection is not formed in the first 10s - it will not be
attempted anymore. And this introduces non-deterministic failures in the
tests.