Stabilize DNS queries in natlab tests
test_dns.py had abstracted calls to nslookup to a dedicated function
but that function was not used in test_dns_through_exit.py. That
function is now moved to a dedicated utils file to make it easier
to share. Additionally, the function was made more stable by slightly
increasing how long the test waits for a DNS response, removing the
retry option passed to nslookup to allow it to retry infinitely,
and lastly by setting the internal timeout for nslookup to only a
second, which means it will retry over and over instead of just
waiting.