wg: Fix deadlock on interface callbacks when stopping
There are a couple of possible deadlocks due to the use of mutexes
when calling CancelMibChangeNotify2(). There is no need to lock
those mutexes inside the callbacks so this commit removes it.

There is a second issue to be solved that route and interface callbacks
in mtumonitor share the same state. It should be cloned per callback or
guarded by a separate mutex shared between these two callbacks.
It will be fixed in a separate PR.