Add check for wireguard-go version
Add wg_go_version_x_y_x to the wireguard-go library and
wireguard-go-rust-wrapper. It us used to detect if we are
linking with the correct version of wireguard-go.
It is added because we've already detected a situation in the past where
an old version of wireguard-go was linked. The bug in the build system
is yet not found nor fixed and this is a temporary step to avoid it in
the future