Create uniffi-compatible versions of FFI interface code
Changing to UniFFI allows us to make changes to our FFI code
that would be incompatible with SWIG, and generally clean up
the interface. To make sure everything builds and runs until
the migration is complete, this commit introduces cleaned up
copies of the existing FFI code. The duplication will not be
long-lived