Add FFI functions to craete Features and Config
The FFI functions for Telio currently take strongly typed objects, but
consumers might have a platform-specific representation of those types
and would have an easier time passing json-encoded strings. To make it
easier for those consumers, this commit introduces functions to create
objects from such strings. At this time we only see a need to do that
for Features and Config, so the conversion functions have only been
implemented for those types