Fix wireguard-go linking issue
When #[link] attribute is used cargo and sccache are not aware of this
dependency. Because of that old version of wireguard-go-rust-wrapper can
be fetched from cache. It is a problem because this fetched .rlib contains
wireguard-go static lib embedded inside (which might be out of date).
Instead lets use cargo:rustc-link-lib command. This way cargo is aware
of this dependency and sccache will take wireguard-go.lib as part of
dependency hash.
This commit also removes safety check put in place before the rootcause
for linking with old version of wireguard-go was found.