nurse: Skip qos analytics ping for unconnected peers
Peers are being pinged for qos analytics regardless of their
connected state. This creates a scenario where resources are being
wasted to ping these peers even though we know they are not connected.

This commit skips pinging these peers and opted to don't touch any
data as analytics can compute if the node is connected or not by
analysing the connection time and the heartbeat interval.

It also cleans some of the logs and increase their level where
appropriate.