Temporary fix dependencies issue
We are using an outdated library in telio-traversal called sm
It uses some old versions of quote, proc-macro2 and syn.
When the moose dependency is injected by the build script it tries to
use these old versions of quote, proc-macro2 and syn. A cargo update
should fix it for now, but it's prone to errors, so we should find an
alternative solution for the sm library.