Add synchronization login to telio-proxy
We were getting the wrong endpoint map from the proxy when trying to
configure the wg controller. There was recently some error handling
added which made the problem visible, but the problem was there for
longer. The root of the issue was that if there was a socket error,
StateIngress would swallow the error while StateEgress would actually
handle it, yet when we were fetching the endpoint map for the wg
controller, we would get it from StateIngress. As a result, if there
was a socket error, ingress and egress could be out of sync, and we
would get the wrong values passed to the wg controller. As a fix,
there is now synchronization logic added for ingress and egress,
and the endpoint map is fetched from egress