Make initial_heartbeat_interval to default at 5 minutes
initial_heartbeat_interval represents the time between the start of
libtelio and the first analytics event. Until now, when it was missing
from the config, it defaulted to None and the first event was generated
after heartbeat_interval time which by default is 1 hour.
It was requested that we use a default value of 5 minutes for
initial_heartbeat_interval.