Improve proxy endpoint synchronizaiton
There was recently a commit that introduced synchronization for the
endpoint maps of StateIngress and StateEgress, which fixed an issue
where StateEgress would update it's endpoint map on socket error, but
StateIngress would be used as the source of truth, leading to an
incorrect endpoint map being used. This commit improves that work
in a few ways. It:
- Adds a test to check that the scenario works
- Makes wg_controller know about both the old endpoint and the new
endpoint
- Resends the data to make libtelio adapt to endpoint changes quicker