Fix wireguard go oscillation
Wireguard-Go works in non standart way on windows, instead of having
one listening socket that is used for all peers, it has two.

One is bound to default interface, to avoid adding routing rules
for all posible endpoints of peers. And other one that is created
on localhost address, and is used to send traffic to peers with
localhost endpoints. These localhost endpoints are used to proxy
wireguard trafic via relay.

Previously we would only report localhost socket's port, that is
invalid for direct connection, leading to failed direct upgrades.

So now we report external port via uapi, and have additional method
to get local one to be used in telio-proxy.