Implement nat state reporting
For a given pair of peers, only peer with a public key that is smaller
of the two, will report events. This means that potentially, the peer which
was requested to upgrade will be reporting and is a reason why the
endpoint types are now sent in the upgrade message.