Workaround for docker bug
https://github.com/moby/moby/issues/47619
Because of this bug eth0 is not created before sysctl are being set so
it is not possible to set any syscts related to the specific
interfaces.
We can create a workaround by not setting eth0. Instead we can set
default so eth0 will have this value assigned later when created.