Add multi-lang code support for rustdoc
We want to have code examples for multiple languages in our docs
to make it more relevant for our known integrators. With vanilla
rustdoc, you would have to show all languages at the same time,
making the page enormous and complicated to read. This commit adds
a language selector as a custom webcomponent so that the rendered page
shows a language selector and only code for the selected language is
shown.