Add initial integration docs
So far, the integration docs is mostly a draft. It is a mostly 1:1
mapping of existing integration docs but for multiple relevant
languages instead of just C. The code examples at this point should
be mostly correct but might be wrong in some places, which will be
fixed in future work.