Use NWPathMonitor to get interfaces in order of preference for socket protector on apple platforms
Using only interfaces with the router property is incorrect since when using
static IPs it's possible to have an interface without router property. Additionally,
interfaces returned by the dynamic store are not in any particular order. After this,
change, we use the interfaces collected from the NWPathMonitor. Those are ordered in
the same way as the OS would prefer them.