PQ: `AuthTag` in the GET KEY request and improve REKEY
It was decided that the current approach for fetching keys
is susceptible to DDoS attack, since the attacker can push
GET KEY request with the public key of a victim to block
his VPN connection for 3 minutes.
With REKEY the current approach with random preshared key is flawed
since the new key is send in plain text (although its
protected by the outter tunnel). The new approach reuses
the Kyber KEM to establish a new preshared key.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>