Fix package-aar and upload-maven scripts
See LLT-5008.

`waitFor()` does not ensure that process will run till compleation, and
cargo metadata method takes a long while to collect info.

From https://docs.groovy-lang.org/latest/html/groovy-jdk/java/lang/Process.html

Gets the output and error streams from a process and reads them to keep
the process from blocking due to a full output buffer. The processed
stream data is appended to the supplied OutputStream. For this, two
Threads are started, so this method will return immediately. The threads
will not be join()ed, even if waitFor() is called. To wait for the
output to be fully consumed call waitForProcessOutput().

Switched to `waitForProcessOutput`.