Use structured events in FFI event callback
Now that we have the type safety in our FFI layer that UniFFI provides
there is no need to keep our FFI event callback dealing with strings.
Natlab still relies on its own types and strings to work and changing
that would be a big piece of work so for now, only a serialization
function is created to keep natlab working. A task for miigrating
natlab to the generated types has been create and  will be done later