Telio-starcast nat
Starcast needs to nat messages comming from the transport socket in
order to know to which peer the response to the multicast message
should be sent to.

There are multiple ways to implement the mapping to the port.
The full-cone approach was chosen because in case of the multicast
the source ip of the response packet will not match the destination
address of the multicast that has opened the mapping so the full-cone is
actually the only option.

In case of the mapping collision the next port number is chosen.