Make event body non-optional
So far, the body of the event we send has been an Option, event though
it should never be. The reason was to use a sort of builder pattern
directly on the events, but this introduces some extra steps for apps
where they have to check that the event, that should never be None,
is none or not before they proceed. To make this easier for them,
we can make a proper builder pattern implementation for events instead