WIP Skeleton batcher
This is highly work in progress and just bare bones.

What's present in this commit:
* batcher crate with Batcher that contains
and add() method that returns a channel to call site.
The channel is resolved each time batcher thinks it's a good time.

Batcher's logic is very straightforward: current time is subtracted
from add time and compared to a threshold which now is hardcoded
as half the period. This seems to work for now.

Session keeper is the sole user of Batcher at the moment. Session
keeper is responsible for direct connections so it's pretty nice area to start.

* NatLab tests for batching. Those are pretty naive at this point
however the skeleton is there: uses tcpdump to capture traffic
on target machines and later on analyzes the traffic. Histogram is generated
and U shape is expected though not enforced yet.

What's not done and missing:
- wg_controller has '|| true' hardcoded in order to add peers to the SessionKeeper,
no idea why it suddenly stopped working.
- multiple uses of unwrap()
- Batcher and SessionKeeper are not properly stopped. This is harder
than it looks as other elements have Arc<T> types meanwhile Batcher
and SessionKeeper need to have Arc<Mutex<T>> due to being modifyable.
The proper way to do that is to wrap both of them into an option
but that get's ugly real fast.
- NatLab doesn't have proper batching tests yet and multiple peers
need to be present at different times in order to check if batching really does
happen.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>