Batcher introduction to SessionKeeper
SessionKeeper was owning it's tasks that used to ping the peers
when added via RepeatedAction.

This commit introduces batcher crate that is a simple wrapper
for variable intervals and has an api for adding an item
ar receiving a Receiver<> channel. Batcher will then select
the best time to emit to the channel.

This mechanism now replaces SessionKeeper's internals
and instead of owning pinging tasks, now it owns channels and
listen for them. Once data is received - SessionKeeper then pings the
peer.

Batcher works by sleeping until the interval of the task
and upon any task wake-up it then checks if any of the tasks
fall into the threshold and if so all are signalled.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>