Add downgrade logic using link detection
The old downgrade logic was triggered if we did not receive anything for
three direct keepalive period.

For battery improvement purpouses we want to increase these keepalive
periods from 5 seconds to somewhere around 120 +- 5 seconds, therefore
the old mechanism for downgrade won't work anymore.

The new solution uses the link detection's reports to trigger the downgrades.
If the link_state of a peer changes to Down and the peer has a direct
connection, we will trigger the downgrade.

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>