Improve docker port mapping
The port mapping is necessary for natlab to be run with macos or
windows as a host, but is not necessary on linux and sometimes breaks
our CI pipeline due to the port being busy on the host. Natlab already
discovers the port mapping fully from the docker information so we
don't need to have the same port on both container and host. So a
better option is to just bind a port in the container and let docker
choose an available port on the host. Additionally, by using
non-ephemeral ports we improve it even further