Let Pyro5 select port automatically
We occasionally have problems in CI that the socket we ask Pyro5 to
bind is busy, causing the pipeline to fail even though nothing is
wrong with the code. Pyro5 can automatically select an available port
and we can query the containers for the port bound by the script that
is starting the Pyro5 instance, so doing that instead of assigning a
random port should improve the situation.