Batcher introduction
Introduced Batcher entity that
controls when signals should be emitted.
There's only one call site at the moment
which is direct peer keepalives.

For safety the old method of using RepeatedActions
is left intact and governed by nat-lab tests by providing,
permutations with batcher feature config disabled(missing)
and with threshold of zero, meaning it should work the same
as with RepeatedActions.

SessionKeeper is reused as it only did pings, no need to copy it.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>