Disable WG-persistent-keepalive for direct peers
Direct peers do not need additional WG persistent-keepalive
presence because it's already being done by SessionKeeper.

Both are good however only one is necessary. Using
SessionKeeper is a bit better as it doesn't depend on
the WireGuard backend as there might be differences,
for example some may not emit if there as data exchange.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>