Make backup of events db in natlab
For our tests, we were copying the original events db from the
container to the host when checking for events. This way of doing it
wouldn't take into account in-progress writes or other reasons why
the db might be locked. As a way to get around it, we now let sqlite
create a backup first, which does respect the db potentially being
locked, and then we copy the backup