Remove fingerprint from nurse feature config
The fingerprint can be fetched from moose, so we don't need to specify
it in the feature config anymore. On top of that, removing it from the
feature config means that we can have a default nurse config, which
simplifies things for apps. Therefore, the nurse config is now active
by default, instead of inactive, but can still be deactivated