Rework unit tests for feature config
The feature config didn't have a very good test suite, with some tests
testing that serde works correctly, which we doesn't need to do, and
some tests just not making much sense. For example, having a test
to check that we can deserialize with all custom values or a single
custom value and the rest be defaults doesn't make sense. It's better
to have one test where all fields are custom, and then test what
happens when no values are supplied.