Make it possible to control using the `hide_ips` if the IP addresses are hidden in logs
IPs should always be hidden in production, but hiding them makes it hard to investigate bugs
found in the nat-lab tests. This is why a flag to control IP hiding has been added. While it
defaults to hiding the IPs, for the nat-lab tests (and only for that case) it defaults to
show IPs by changing pytests default value.

Additionally, as a side effect, the log censoring functionality no longer makes copies of
log message strings when they haven't been modified.