natlab: Trigger garbage collection when dropping node
This test starts one of the nodes without any EP and expects to
accomplish relay path connection. Then it recreates the node to add
the stun server as EP, expecting then a direct path connection.

This test was failing locally because although the telio.Client
instance was dropped it was ran inside the main exit stack,
so not collected by the python's gc while inside the exit stack.
This resulted in a failed second telio.Client init with
a "Address already in use" OS error when attempting the Pyro5
connection.

This commit replaces the setup_mesh() function by individual exit
stacs, so that the dropped node is fully cleaned by the
python's garbage collector before recreating the libtelio remote
connection.