Add batch observing natlab test
`test_batching.py` is added
that contains a star-like configuration of few
meshnet nodes that are all interconnected.

This allows to achieve as much testing as possible for batching
by also testing if things were not broken.

For better batching observability peers are first
connected, then stopped, then one by one, with a delay, started again.

This works by capturing packet captures with tcpdump on all the
nodes as well as gateways and later on running filters and rendering
histograms.

This commit marks the beginning of such an effort.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>