Fix incorrect entries in nat_traversal_conn_info
1. Previusly report_peer_state_direct and report_peer_state_relayed were
triggered by different events. Because of that there was a possibility
for the reorder of those events causing incorrect reporting. Now we
report it always from a single place in the wg_controller so there is no
risk of a race condition.
2. Fix bug in get_current_meshmap_info in heartbeat. This bug caused
last intrnal node and first external node to have the same index