Add network-activity trigger ability for batcher
Add triggering ability to batcher so it could
evaluate deadlines and thresholds on demand instead
on when some signal is awoken.

The approach is simple - any activity on sent
or received data on any peer will trigger the batcher.
It is built on two assumptions:
1) It is believed that triggering on incoming data
would "sync" the batchers on different libtelio instances.
This is because one would receive, send data back, effectively
reseting it's intervals at nearly the same time.
2) receiving data hurts the battery to a similar degree as
sending(practical tests suggest).

Caution: this feature introduces possibility of a meltdown,
similar to TCP over TCP, where one batcher might trigger another,
which in turn has ability to trigger the original one back.
Repeated this would spiral out of control.

For this safeguards are added in the code:
- Minimal interval is 5seconds.
  This is because just recently the only action which
  had this frequency was direct peer keepalives which
  is not defaulted to 10seconds.
- Maximum threshold is half the interval.
- Each action contains timestamp since it was
batched which is then used to rate-limit with 4seconds
hardcoded constant.

Side effects:
consider having 3 interconnected peers: A, B and C.
Peer C is idling and A streams data to B.
Now A and B is each triggered on every packet and in turn
send premature keepalives to C at T_new=T_orig-threshold.

Triggering will happen mostly on wg_consolidate() which happens
every second.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>