Update rust_build_utils ref to fix dylib rpath
When producing dylibs(Apple dynamic libraries) the
*id* of the application is an absolute path which is
relative to build machine.

This causes linker on Apple platforms to fail linking against
this exact dylib and doing `install_name_tool -id @rpath/libtelio.dylib`
is required to fix it.

This commit bumps the ref of rust_build_utils that contains this fix.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>