Call logging callback on a separate thread
To avoid blocking callers, when logging is slow, the call to the logging
callback is moved to the separate therad. Now, logging anything only
causes non blocking try_send that will fail if there is no more space in
the queue of not yet processed log messages.

To have some visibility into how this system performs, count of dropped
logs is printed along side one of the periodic logs of libtelio.