Increase constrains in link state testcases
Previously the link state testcase scenarios were structured like setup,
do some interaction between peers, sleep for a while and check the reported link states.
This approach did not consider the timings of the events.

Proposed solution consists in checking the state constantly during the
scenario, not only at the end of it.

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>